finish run_dataset()
