+++ src/MSA_FET/aligners/__init__.py
@@ -1,7 +1,5 @@
 from .wav2vec2 import Wav2Vec2Aligner
 
-__all__ = ['ALIGNER_MAP']
-
 ALIGNER_MAP = {
     "wav2vec": Wav2Vec2Aligner,
 }
