+++ src/MSA_FET/aligners/wav2vec2.py
@@ -96,3 +96,76 @@ class Wav2Vec2Aligner(BaseAligner):
     
     def get_asr_result(self) -> str:
         return self.asr_text
+
+    # @staticmethod
+    # def do_asr(
+    #     audio_file : str | Path, 
+    #     model_name : str = "jonatasgrosman/wav2vec2-large-xlsr-53-english",
+    #     device : torch.device = torch.device("cpu")
+    # ) -> str:
+    #     try:
+    #         processor = Wav2Vec2Processor.from_pretrained(model_name)
+    #         model = Wav2Vec2ForCTC.from_pretrained(model_name).to(device)
+    #         sample_rate = 16000
+    #         audio, _ = librosa.load(audio_file, sr=sample_rate)
+    #         features = processor(
+    #             audio, 
+    #             sampling_rate=sample_rate,
+    #             return_tensors="pt", 
+    #             padding="longest"
+    #         )
+    #         with torch.no_grad():
+    #             logits = model(features.input_values.to(device)).logits.cpu()[0]
+
+    #         predicted_ids = torch.argmax(logits, dim=-1)
+    #         asr_text = processor.decode(predicted_ids)
+    #         return asr_text
+    #     except Exception as e:
+    #         raise e
+    
+    # @staticmethod
+    # def do_alignment(
+    #     audio_file : str | Path, 
+    #     transcript : str, 
+    #     model_name : str = "jonatasgrosman/wav2vec2-large-xlsr-53-english",
+    #     device : torch.device = torch.device("cpu")
+    # ) -> list[dict]:
+    #     try:
+    #         processor = Wav2Vec2Processor.from_pretrained(model_name)
+    #         tokenizer = Wav2Vec2CTCTokenizer.from_pretrained(model_name)
+    #         model = Wav2Vec2ForCTC.from_pretrained(model_name).to(device)
+    #         sample_rate = 16000
+    #         audio, _ = librosa.load(audio_file, sr=sample_rate)
+    #         features = processor(
+    #             audio, 
+    #             sampling_rate=sample_rate, 
+    #             return_tensors="pt", 
+    #             padding="longest"
+    #         )
+    #         with torch.no_grad():
+    #             logits = model(features.input_values.to(device)).logits.cpu()[0]
+    #             probs = torch.nn.functional.log_softmax(logits,dim=-1)
+
+    #         # Tokenize transcripts
+    #         transcripts = transcript.split(" ")
+    #         vocab = tokenizer.get_vocab()
+    #         inv_vocab = {v:k for k,v in vocab.items()}
+    #         unk_id = vocab["<unk>"]
+    #         tokens = []
+    #         for transcript in transcripts:
+    #             assert len(transcript) > 0
+    #             tok_ids = tokenizer(transcript.lower())['input_ids']
+    #             tok_ids = np.array(tok_ids,dtype=np.int32)
+    #             tokens.append(tok_ids[tok_ids != unk_id])
+            
+    #         # Do align
+    #         char_list = [inv_vocab[i] for i in range(len(inv_vocab))]
+    #         config = ctc_segmentation.CtcSegmentationParameters(char_list=char_list)
+    #         config.index_duration = audio.shape[0] / probs.size()[0] / sample_rate
+            
+    #         ground_truth_mat, utt_begin_indices = ctc_segmentation.prepare_token_list(config, tokens)
+    #         timings, char_probs, _ = ctc_segmentation.ctc_segmentation(config, probs.numpy(), ground_truth_mat)
+    #         segments = ctc_segmentation.determine_utterance_segments(config, utt_begin_indices, char_probs, timings, transcripts)
+    #         return [{"text" : t, "start" : p[0], "end" : p[1], "conf" : np.exp(p[2])} for t,p in zip(transcripts, segments)]
+    #     except Exception as e:
+    #         raise e
